# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: ThoorigAI Infotech Admin Dashboard API
  version: 0.1.0
  description: >-
    App Router HTTP contract. Protected routes require the Supabase cookie session
    and the role permission documented in docs/architecture/role-permissions.md.
    Route handlers and Zod schemas remain authoritative when this reference changes.
servers:
  - url: https://{host}
    variables:
      host:
        default: localhost:3000
security:
  - cookieAuth: []
paths:
...
  /api/payments:
    get:
      operationId: listPayments
      summary: List and search invoices/payments with pagination
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - { $ref: '#/components/parameters/Search' }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ListResult' }
    post:
      operationId: recordPayment
      summary: Record a payment transactionally
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema: { type: string, minLength: 1, maxLength: 200 }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, description: Validated by the payment Zod schema }
      responses:
        '201': { description: Payment, invoice number, and verification code returned }
        '409': { description: Idempotency key reused with a different request }
        '422': { description: Payment exceeds the remaining balance }
...
  /api/certificates:
    get:
      operationId: listCertificates
      summary: List certificates
      responses:
        '200': { $ref: '#/components/responses/ListResult' }
    post:
      operationId: issueCertificate
      summary: Issue a fee-cleared certificate
      responses:
        '201': { description: Certificate and public verification document created }
        '422': { description: Fees remain outstanding }
...
  /api/audit:
    get:
      operationId: listAuditLog
      summary: List immutable audit records (admin only)
      parameters:
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/PageSize' }
        - name: cursor
          in: query
          schema: { type: string }
      responses:
        '200': { $ref: '#/components/responses/ListResult' }
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-<project-ref>-auth-token
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Search:
      name: search
      in: query
      schema: { type: string, maxLength: 100 }
  responses:
    ListResult:
      description: Paginated `{ data, meta }` response
      content:
        application/json:
          schema:
            type: object
            properties:
              data: { type: array, items: { type: object } }
              meta: { type: object }
...